Fix compiler deprecation and exhaustivity warnings
- Renderer.scala (x2), StartupRenderingSpec.scala: explicit .toFloat
  instead of the deprecated implicit Int-to-Float widening
- BenchmarkRunner.scala: Thread.threadId() instead of the deprecated
  Thread.getId()
- WorkspaceTreeLifecycleSpec.scala: collapse the non-Split match arms
  to a single wildcard case so the match is exhaustive against
  WorkspaceNode.DockedSurface

+++ src/main/scala/com/serenity/ui/renderer/Renderer.scala
@@ -2756,7 +2756,7 @@ object Renderer:
             area = TextAreaPx(
               xPx = cellMetrics.toPixelX(rect.x).toFloat,
               yPx = yPx,
-              widthPx = rect.width * cellMetrics.charWidth,
+              widthPx = rect.width.toFloat * cellMetrics.charWidth,
               heightPx = lineHeightPx
             ),
             font = font,
@@ -2848,7 +2848,7 @@ object Renderer:
           area = TextAreaPx(
             xPx = 0.0f,
             yPx = yPx,
-            widthPx = viewportSize.width * cellMetrics.charWidth,
+            widthPx = viewportSize.width.toFloat * cellMetrics.charWidth,
             heightPx = lineHeightPx
           ),
           font = uiFont,
